pyproject.toml
[project]
name = "nlp-ii-practica2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "datasets>=4.4.1",
    # Core Deep Learning (Manteniendo tus versiones superiores)
    "torch>=2.9.1",
    "torchaudio>=2.9.1",
    "torchvision>=0.24.1",
    "transformers>=4.57.3",
    "tokenizers>=0.15.0",
    
    # Kaggle
    "kagglehub>=0.1.0",
    
    # Data Processing
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "openpyxl>=3.1.0",
    "xlrd>=2.0.0",
    
    # Machine Learning
    "scikit-learn>=1.3.0",
    
    # Visualization
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    
    # Utilities
    "tqdm>=4.65.0",
    
    # Vector Database (GPU)
    "faiss-cpu>=1.7.4",
    
    # Jupyter Environment
    "jupyter>=1.0.0",
    "ipywidgets>=8.0.0",
]

[tool.uv.sources]
torch = { index = "pytorch_cu130" }
torchvision = { index = "pytorch_cu130" }
torchaudio = { index = "pytorch_cu130" }

[[tool.uv.index]]
name = "pytorch_cu130"
url = "https://download.pytorch.org/whl/cu130"
explicit = true